+++ src/main/java/com/board/member/domain/auth/TokenProvider.java
@@ -1,9 +1,7 @@
 package com.board.member.domain.auth;
 
-import com.auth0.jwt.interfaces.DecodedJWT;
-
 public interface TokenProvider {
 
     String create(Long memberId);
-    DecodedJWT verifyToken(String token);
+    Long extractMemberId(String token);
 }
