+++ src/main/java/com/board/member/service/auth/AuthService.java
@@ -1,10 +1,5 @@
 package com.board.member.service.auth;
 
-import com.auth0.jwt.exceptions.JWTDecodeException;
-import com.auth0.jwt.exceptions.TokenExpiredException;
-import com.auth0.jwt.interfaces.DecodedJWT;
-import com.board.global.resolver.exception.TokenExpirationException;
-import com.board.global.resolver.exception.TokenVerificationException;
 import com.board.member.controller.auth.dto.request.LoginRequest;
 import com.board.member.controller.auth.dto.request.SignUpRequest;
 import com.board.member.controller.auth.dto.response.SignUpResponse;
@@ -13,8 +8,6 @@ import com.board.member.domain.member.Member;
 import com.board.member.exception.MemberErrorCode;
 import com.board.member.exception.MemberException;
 import com.board.member.repository.MemberRepository;
-import com.board.global.resolver.exception.TokenInvalidException;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -42,26 +35,6 @@ public class AuthService {
         member.checkPassword(request.password());
 
         return tokenProvider.create(member.getId());
-    }
-
-    public Long verifyAndExtractToken(String token) {
-        try {
-            return extractToken(token);
-        } catch (JWTDecodeException e) {
-            throw new TokenVerificationException();
-        } catch (TokenExpiredException e) {
-            throw new TokenExpirationException();
-        }
-    }
-
-    private Long extractToken(String token) {
-        return verifyToken(token).getClaim("memberId")
-                .asLong();
-    }
-
-    private DecodedJWT verifyToken(String token) {
-        return Optional.of(tokenProvider.verifyToken(token))
-                .orElseThrow(TokenInvalidException::new);
     }
 
     private Member findMemberByLoginId(String loginId) {
